# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rinawarp API
  version: 1.0.0
  description: API for the Rinawarp AI-powered terminal platform
  contact:
    name: Rinawarp Support
    email: support@rinawarp.com
    url: https://rinawarp.com/support
servers:
  - url: https://api.rinawarp.com/v1
    description: Production server
  - url: https://staging-api.rinawarp.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development
security:
  - bearerAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
                - refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: Invalid refresh token
...